Fix remaining compilation errors
1. Updated ReloadableRegistry import path:
   - Changed from uk.joshiejack.penguinlib.data.database.ReloadableRegistry
   - To uk.joshiejack.penguinlib.util.registry.ReloadableRegistry

2. Fixed Page.CODEC for polymorphic feature deserialization:
   - Removed reference to abstract FeatureProvider.CODEC
   - Created FEATURE_CODEC using registry dispatch pattern
   - Uses EnchiridionRegistries.Features.FEATURE for polymorphic codec resolution

3. Disabled legacy GSON adapter for FeatureProvider:
   - FeatureProvider is now abstract and cannot be instantiated
   - Commented out FeatureProvider.class registration in GsonHelper
   - Updated FeatureAbstractAdapter to throw UnsupportedOperationException
   - Marked adapter as @Deprecated with documentation
   - Modern serialization uses Codec-based approach

4. Fixed IRecipeHandler implementation signatures:
   - RecipeHandlerFurnace: Changed getHeight/getWidth to return int (was double)
   - RecipeHandlerRecipeBase: Changed getHeight/getWidth to return int (was double)
   - Now matches interface declaration: int getHeight(int width) and int getWidth(int width)

All changes maintain compatibility with the new FeatureProvider architecture
where features extend FeatureProvider directly and use Codec-based serialization.

+++ src/main/java/joshie/enchiridion/gui/book/features/recipe/RecipeHandlerFurnace.java
@@ -43,12 +43,12 @@ public class RecipeHandlerFurnace extends RecipeHandlerBase {
     }
 
     @Override
-    public double getHeight(double width) {
-        return width * 1.1D;
+    public int getHeight(int width) {
+        return (int) (width * 1.1D);
     }
 
     @Override
-    public double getWidth(double width) {
+    public int getWidth(int width) {
         return width;
     }
